requirements.txt
requests==2.26.0
spotipy==2.18.0
qrcode==7.3.1
reportlab==3.6.5
python-dotenv==0.19.1
pandas==1.3.3
